2x2 matrix db with :matrix, :row and :cell. only does render, no mutate.
Here I'm trying to image what a db of the matrix might be like. I don't think it's right. Extracting the data from the db is too complicated. It may be simpler to represent the cells as a series and not include the row and matrix info. Or maybe just include the dimensions of the matrix, plus the cells.